docs: annotate unclear code purpose
Nightshift-Task: why-annotator

Nightshift-Ref: https://github.com/marcus/nightshift
