+++ docs/README.md
@@ -1,1 +1,3 @@
 This directory contains the documentation about this extension.
+
+- [WHY_ANNOTATOR.md](./WHY_ANNOTATOR.md): purpose notes for runtime hotspots with non-obvious behavior.
